ecommerce-pipeline-project/requirements.txt
pandas>=2.0.0
rich>=13.0.0
apache-airflow>=2.7.0
apache-airflow-providers-common-sql>=1.7.0
